requirements.txt
streamlit>=1.28.0
googlesearch-python>=1.2.3
pandas>=2.0.0
openpyxl>=3.1.0
